requirement.txt
fastapi==0.111.0
uvicorn==0.30.1
pymysql==1.1.0
SQLAlchemy==2.0.20
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==2.7.1
streamlit==1.35.0
requests==2.32.3